# optimization: buffer reads(23%)/writes(290%) in [undo]block [de]serialization, 6% faster IBD