--- a/src/streams.h
+++ b/src/streams.h
@@ -15,6 +15,7 @@
 #include <assert.h>
 #include <cstddef>
 #include <cstdio>
+#include <flatfile.h>
 #include <ios>
 #include <limits>
 #include <optional>
@@ -23,6 +24,7 @@
 #include <string>
 #include <utility>
 #include <vector>
+#include <util/check.h>
 
 namespace util {
 inline void Xor(Span<std::byte> write, Span<const std::byte> key, size_t key_offset = 0)
@@ -612,6 +614,42 @@ public:
             if (buf_offset >= vchBuf.size()) buf_offset = 0;
         }
     }
+};
+
+class BufferedReadOnlyFile
+{
+    AutoFile m_src;
+    std::vector<std::byte> m_buf;
+    size_t m_buf_start{0}, m_buf_end{0};
+
+public:
+    explicit BufferedReadOnlyFile(const FlatFileSeq& block_file_seq,
+                                  const FlatFilePos& pos,
+                                  const std::vector<std::byte>& m_xor_key,
+                                  const size_t buf_size = 16 << 10)
+        : m_src{block_file_seq.Open(pos, /*read_only=*/true), m_xor_key},
+          m_buf{buf_size} {}
+
+    void read(Span<std::byte> dst)
+    {
+        if (m_buf_start < m_buf_end) {
+            const size_t chunk = Assert(std::min(dst.size(), m_buf_end - m_buf_start));
+            std::memcpy(dst.data(), m_buf.data() + m_buf_start, chunk);
+            m_buf_start += chunk;
+            dst = dst.subspan(chunk);
+        }
+        if (!dst.empty()) {
+            Assume(m_buf_start == m_buf_end);
+            m_src.read(dst);
+
+            m_buf_start = 0;
+            m_buf_end = m_src.detail_fread(m_buf);
+        }
+    }
+
+    bool IsNull() const { return m_src.IsNull(); }
+
+    template <typename T> void operator>>(T&& obj) { ::Unserialize(*this, obj); }
 };
 
 #endif // BITCOIN_STREAMS_H