@l0rinc 2025-01-23
You may be right, though I don't yet see how this memory saving option keeps crashing after the rebase (it was passing before), while the [memory hogging](https://github.com/bitcoin/bitcoin/pull/31551) alternative is passing.
https://github.com/bitcoin-dev-tools/benchcoin/pull/123#issuecomment-2610155151 shows the before/after memory to be the same.

I suspected a stuck file or process or infinite loop...

> asan CI max memory usage

I don't yet know how to run that on a Mac (and my linux boxes are running IBDs for now).
I'll push a few more rounds here to try to understand which part is causing the problem - and why only for `ASan + LSan + UBSan + integer`, if they won't reveal anything, I'll investigate how to replicate the CI behavior locally.