@l0rinc 2025-01-24
It looks like the problem was with `BufferedWriteOnlyFile` - if needed, I will investigate further.

Given that there is an [alternative PR to this](https://github.com/bitcoin/bitcoin/pull/31551) (reading the whole block into memory (instead of smaller chunks) - which we're already doing in other places, so shouldn't be controversial), I'll close this PR. If the memory allocations come up in that, I'll consider reopening.